Update overview.adoc
Removing Spot agents as a default partition on the cluster and adding a note indicating that this architecture could be achieved as an advanced use case. Looking for feedback on if this is the best approach. The Spot EC2 service is linked in the 'AWS Services' section at the end of the guide.