removing references to Spot instances
Also: changing wording on line 16 as the sentence was incomplete. I'm curious about what to do with lines 46-54 that seem to be specific to spot interruption events. It should be removed from this section since we won't be deploying a spot partition by default, however, IMO that information is nice to have for those who opt-in for spot instances.